Updated dir structure
