+++ Tools/SeeDot/seedot/ast/printAST.py
@@ -5,10 +5,10 @@
 PrintAST can be used to print the generated AST for a given SeeDot program.
 '''
 
-from antlr.seedotParser import seedotParser as SeeDotParser
+from seedot.antlr.seedotParser import seedotParser as SeeDotParser
 
-import ast.ast as AST
-from ast.astVisitor import ASTVisitor
+import seedot.ast.ast as AST
+from seedot.ast.astVisitor import ASTVisitor
 
 indent = "  "
