+++ Tools/SeeDot/seedot/converter/protonn.py
@@ -3,7 +3,7 @@
 
 import os
 
-from converter.util import *
+from seedot.converter.util import *
 
 # Class to read ProtoNN model dumps and generate input files (C header file and the compiler input)
 # The two classes ProtonnFixed and ProtonnFloat are for generating fixed model and float model respectively
