+++ Tools/SeeDot/seedot/test.py
@@ -1,0 +1,4 @@
+
+import numpy as np
+
+train = np.load('SeeDot/datasets/usps-binary/train.npy')
