+++ Tools/SeeDot/seedot/writer.py
